+++ src/pages/Tenant/helpers.ts
@@ -1,22 +1,8 @@
+/* eslint-disable indent */
 import { routes } from 'utils/routes'
+import { PagesConfig, PagesGroup } from './types'
 
 const defaults = [routes.send, routes.faucet, routes.staking, routes.vesting, routes.launchpad]
-
-interface PagesConfig {
-  dex: boolean
-  kyc: boolean
-  lbp: boolean
-  lbpAdmin: boolean
-  offer: boolean
-  issuance: boolean
-  payout: boolean
-  securityTokens: boolean
-  admin: boolean
-}
-
-interface PagesGroup {
-  [key: string]: string[]
-}
 
 export const pagesGroup: PagesGroup = {
   admin: [
@@ -57,7 +43,7 @@ export const pagesGroup: PagesGroup = {
     routes.payoutHistory,
     routes.payoutEvent,
     routes.tokenManager('my-tokens', null),
-    routes.tokenManager()
+    routes.tokenManager(),
   ],
   defaults,
 }
@@ -128,4 +114,106 @@ export function checkExistInPageGroup(pagesSource: string) {
   }
 
   return pages
+}
+
+function checkObjectEmpty(obj: any) {
+  return Object.keys(obj).length === 0
+}
+
+export function generateTenantSubmitPayload(values: any) {
+  const socialLinks = {} as any
+  const footerConfig = {} as any
+  values?.telegram && (socialLinks['telegram'] = values?.telegram)
+  values?.linkedin && (socialLinks['linkedin'] = values?.linkedin)
+  values?.youtube && (socialLinks['youtube'] = values?.youtube)
+  values?.twitter && (socialLinks['twitter'] = values?.twitter)
+  !checkObjectEmpty(socialLinks) && (footerConfig['socialLinks'] = socialLinks)
+  values?.termLink && (footerConfig['termsLink'] = values?.termLink)
+  values?.policyLink && (footerConfig['policyLink'] = values?.policyLink)
+  values?.block1 && (footerConfig['block1'] = values?.block1)
+  values?.block2 && (footerConfig['block2'] = values?.block2)
+  values?.block3 && (footerConfig['block3'] = values?.block3)
+
+  const payload = {
+    name: values?.name,
+    title: values?.title,
+    appUrl: values?.appUrl,
+    description: values?.description,
+    bannerImageUrl: values?.bannerImageUrl,
+    pages: values?.pages ? getActiveRoutes(values.pages) : null,
+    chartsUrl: values?.chartsUrl,
+    enableLbp: values?.enableLbp,
+    defaultUrl: values?.defaultUrl,
+    faviconUrl: values?.faviconUrl,
+    logoUrl: values?.logoUrl,
+    enableFeaturedSecurityVaults: values?.enableFeaturedSecurityVaults ?? false,
+    colors: values?.colorButtonPrimary
+      ? JSON.stringify({
+          button: { primary: values?.colorButtonPrimary },
+        })
+      : '',
+    customStyles: '{}',
+    supportEmail: values?.supportEmail,
+    isIxSwap: values?.isIxSwap ?? false,
+    tokens: values?.tokens ? JSON.stringify(values.tokens) : null,
+    domain: values?.domain,
+    enableLaunchpadBanner: values?.enableLaunchpadBanner ?? false,
+    launchpadBannerTitle: values?.launchpadBannerTitle,
+    launchpadBannerInfoRedirectTitle: values?.launchpadBannerInfoRedirectTitle,
+    launchpadBannerInfoRedirectUrl: values?.launchpadBannerInfoRedirectUrl,
+    kycSuccessRedirectUrl: values?.kycSuccessRedirectUrl,
+    kycCancelRedirectUrl: values?.kycCancelRedirectUrl,
+    footerConfig: !checkObjectEmpty(footerConfig) ? JSON.stringify(footerConfig) : '',
+  }
+
+  return payload
+}
+
+export const pages = {
+  dex: false,
+  kyc: false,
+  lbp: false,
+  lbpAdmin: false,
+  offer: false,
+  issuance: false,
+  payout: false,
+  securityTokens: false,
+  admin: false,
+}
+
+export function setFieldsValue(setFieldValue: any, data: any) {
+  setFieldValue('name', data.name)
+  setFieldValue('title', data.title)
+  setFieldValue('domain', data.domain)
+  setFieldValue('appUrl', data.appUrl)
+  setFieldValue('description', data.description)
+  setFieldValue('bannerImageUrl', data.bannerImageUrl)
+  setFieldValue('isIxSwap', data.isIxSwap)
+  setFieldValue('enableLbp', data.enableLbp)
+  setFieldValue('enableFeaturedSecurityVaults', data.enableFeaturedSecurityVaults)
+  setFieldValue('chartsUrl', data.chartsUrl)
+  setFieldValue('defaultUrl', data.defaultUrl)
+  setFieldValue('tokens', data?.tokens ? JSON.parse(data.tokens) : [])
+  setFieldValue('logoUrl', data.logoUrl)
+  setFieldValue('faviconUrl', data.faviconUrl)
+  setFieldValue('supportEmail', data.supportEmail)
+  setFieldValue('enableLaunchpadBanner', data.enableLaunchpadBanner)
+  setFieldValue('launchpadBannerTitle', data.launchpadBannerTitle)
+  setFieldValue('launchpadBannerInfoRedirectTitle', data.launchpadBannerInfoRedirectTitle)
+  setFieldValue('launchpadBannerInfoRedirectUrl', data.launchpadBannerInfoRedirectUrl)
+  setFieldValue('kycSuccessRedirectUrl', data.kycSuccessRedirectUrl)
+  setFieldValue('kycCancelRedirectUrl', data.kycCancelRedirectUrl)
+  const footerConfig = data.footerConfig ? JSON.parse(data.footerConfig) : null
+  setFieldValue('termLink', footerConfig?.termsLink)
+  setFieldValue('policyLink', footerConfig?.policyLink)
+  setFieldValue('block1', footerConfig?.block1)
+  setFieldValue('block2', footerConfig?.block2)
+  setFieldValue('block3', footerConfig?.block3)
+  setFieldValue('telegram', footerConfig?.socialLinks?.telegram)
+  setFieldValue('linkedin', footerConfig?.socialLinks?.linkedin)
+  setFieldValue('youtube', footerConfig?.socialLinks?.youtube)
+  setFieldValue('twitter', footerConfig?.socialLinks?.twitter)
+  const colors = data.colors ? JSON.parse(data.colors) : null
+  setFieldValue('colorButtonPrimary', colors?.button?.primary)
+  setFieldValue('pages', data?.pages ? checkExistInPageGroup(data.pages) : pages)
 }
